+++ src/oidc-agent/oidcd/oidcd_handler.c
@@ -42,7 +42,7 @@
 
 void initAuthCodeFlow(struct oidc_account* account, struct ipcPipe pipes,
                       const char* info, const char* nowebserver_str,
-                      const char* noscheme_str, const int only_at,
+                      const char* noscheme_str, const unsigned char only_at,
                       const struct arguments* arguments) {
   if (arguments->no_webserver || strToInt(nowebserver_str)) {
     account_setNoWebServer(account);
@@ -67,7 +67,7 @@ void initAuthCodeFlow(struct oidc_account* account, struct ipcPipe pipes,
   char* code_verifier = secAlloc(CODE_VERIFIER_LEN + 1);
   randomFillBase64UrlSafe(code_verifier, CODE_VERIFIER_LEN);
 
-  char* uri = buildCodeFlowUri(account, state_ptr, &code_verifier);
+  char* uri = buildCodeFlowUri(account, state_ptr, &code_verifier, only_at);
   if (uri == NULL) {
     ipc_writeOidcErrnoToPipe(pipes);
     secFree(code_verifier);
@@ -87,14 +87,6 @@ void initAuthCodeFlow(struct oidc_account* account, struct ipcPipe pipes,
                     *state_ptr);
   }
   secFree(uri);
-}
-
-char* removeScope(char* scopes, char* rem) {
-  scopes = strremove(scopes, rem);
-  scopes = strelimIfAfter(
-      scopes, ' ',
-      ' ');  // strremove leaves a doubled space; this call removes one of it.
-  return scopes;
 }
 
 void _handleGenFlows(struct ipcPipe pipes, struct oidc_account* account,
@@ -237,8 +229,8 @@ void oidcd_handleGen(struct ipcPipe pipes, const char* account_json,
   }
 
   const int only_at = strToInt(only_at_str);
-  char* scope = only_at ? removeScope(oidc_strcopy(account_getScope(account)),
-                                      OIDC_SCOPE_OFFLINE_ACCESS)
+  char* scope = only_at ? _removeScope(oidc_strcopy(account_getScope(account)),
+                                       OIDC_SCOPE_OFFLINE_ACCESS)
                         : NULL;
 
   _handleGenFlows(pipes, account, flow, scope, only_at, nowebserver_str,
@@ -781,7 +773,8 @@ void oidcd_handleRegister(struct ipcPipe pipes, const char* account_json,
   }
   char* scopes = getJSONValue(json_res, OIDC_KEY_SCOPE);
   secFreeJson(json_res);
-  if (!strSubStringCase(scopes, OIDC_SCOPE_OPENID) ||
+  if ((!strSubStringCase(scopes, OIDC_SCOPE_OPENID) &&
+       !account_getIsOAuth2(account)) ||
       !strSubStringCase(scopes, OIDC_SCOPE_OFFLINE_ACCESS)) {
     // did not get all scopes necessary for oidc-agent
     oidc_errno = OIDC_EUNSCOPE;
@@ -1018,13 +1011,15 @@ void oidcd_handleLock(struct ipcPipe pipes, const char* password, int _lock) {
 }
 
 void oidcd_handleScopes(struct ipcPipe pipes, const char* issuer_url,
-                        const char* cert_path) {
-  if (issuer_url == NULL) {
-    ipc_writeToPipe(pipes, RESPONSE_ERROR, "Bad Request: issuer url not given");
+                        const char* config_endpoint, const char* cert_path) {
+  if (issuer_url == NULL && config_endpoint == NULL) {
+    ipc_writeToPipe(
+        pipes, RESPONSE_ERROR,
+        "Bad Request: issuer url or configuration endpoint must be given");
     return;
   }
   agent_log(DEBUG, "Handle scope lookup request for %s", issuer_url);
-  char* scopes = getScopesSupportedFor(issuer_url, cert_path);
+  char* scopes = getScopesSupportedFor(issuer_url, config_endpoint, cert_path);
   if (scopes == NULL) {
     ipc_writeOidcErrnoToPipe(pipes);
     return;